Split full node tests directory for parallelism (#10108)
* Split full node tests directory for parallelism

* reduce full node job timeout to 40, duplicate config to new directory

* rebuild workflows

* add missing __init__.py

* update some imports

* oops

* isort

* Revert "isort"

This reverts commit 03d836034262ed5d86a71edb4a1a06fb19fb6014.

* update isort exclude list
