+++ .github/workflows/build-test-macos-core-full_node-stores.yml
@@ -1,7 +1,7 @@
 #
 # THIS FILE IS GENERATED. SEE https://github.com/Chia-Network/chia-blockchain/tree/main/tests#readme
 #
-name: Ubuntu core-full_node Test
+name: MacOS core-full_node-stores Tests
 
 on:
   push:
@@ -20,15 +20,15 @@ concurrency:
 
 jobs:
   build:
-    name: Ubuntu core-full_node Test
+    name: MacOS core-full_node-stores Tests
     runs-on: ${{ matrix.os }}
     timeout-minutes: 40
     strategy:
       fail-fast: false
       max-parallel: 4
       matrix:
-        python-version: [3.7, 3.8, 3.9]
-        os: [ubuntu-latest]
+        python-version: [3.8, 3.9]
+        os: [macOS-latest]
 
     steps:
     - name: Checkout Code
@@ -41,13 +41,12 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
 
-    - name: Cache npm
-      uses: actions/cache@v2.1.6
-      with:
-        path: ~/.npm
-        key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
-        restore-keys: |
-          ${{ runner.os }}-node-
+    - name: Create keychain for CI use
+      run: |
+        security create-keychain -p foo chiachain
+        security default-keychain -s chiachain
+        security unlock-keychain -p foo chiachain
+        security set-keychain-settings -t 7200 -u chiachain
 
     - name: Get pip cache dir
       id: pip-cache
@@ -57,6 +56,7 @@ jobs:
     - name: Cache pip
       uses: actions/cache@v2.1.6
       with:
+        # Note that new runners may break this https://github.com/actions/cache/issues/292
         path: ${{ steps.pip-cache.outputs.dir }}
         key: ${{ runner.os }}-pip-${{ hashFiles('**/setup.py') }}
         restore-keys: |
@@ -77,17 +77,12 @@ jobs:
         echo "$HOME/.chia"
         ls -al $HOME/.chia
 
-    - name: Install ubuntu dependencies
-      run: |
-        sudo apt-get install software-properties-common
-        sudo add-apt-repository ppa:deadsnakes/ppa
-        sudo apt-get update
-        sudo apt-get install python${{ matrix.python-version }}-venv python${{ matrix.python-version }}-distutils git -y
-
     - name: Run install script
       env:
         INSTALL_PYTHON_VERSION: ${{ matrix.python-version }}
+        BUILD_VDF_CLIENT: "N"
       run: |
+        brew install boost
         sh install.sh -d
 
     - name: Install timelord
@@ -96,17 +91,10 @@ jobs:
         sh install-timelord.sh
         ./vdf_bench square_asm 400000
 
-    - name: Test core-full_node code with pytest
+    - name: Test core-full_node-stores code with pytest
       run: |
         . ./activate
-        ./venv/bin/py.test tests/core/full_node/test_*.py -s -v --durations 0
-
-
-    - name: Check resource usage
-      run: |
-        sqlite3 -readonly -separator " " .pymon "select item,cpu_usage,total_time,mem_usage from TEST_METRICS order by mem_usage desc;" >metrics.out
-        ./tests/check_pytest_monitor_output.py <metrics.out
-
+        ./venv/bin/py.test tests/core/full_node/stores/test_*.py -s -v --durations 0
 #
 # THIS FILE IS GENERATED. SEE https://github.com/Chia-Network/chia-blockchain/tree/main/tests#readme
 #
