+++ .github/workflows/build-test-macos-core-full_node.yml
@@ -22,7 +22,7 @@ jobs:
   build:
     name: MacOS core-full_node Tests
     runs-on: ${{ matrix.os }}
-    timeout-minutes: 80
+    timeout-minutes: 40
     strategy:
       fail-fast: false
       max-parallel: 4
