+++ .isort.cfg
@@ -23,7 +23,6 @@ extend_skip=
     chia/cmds/plotnft_funcs.py
     chia/cmds/plotnft.py
     chia/cmds/plotters.py
-    chia/cmds/seeder.py
     chia/cmds/show.py
     chia/cmds/start_funcs.py
     chia/cmds/start.py
@@ -72,8 +71,6 @@ extend_skip=
     chia/seeder/crawler.py
     chia/seeder/crawl_store.py
     chia/seeder/dns_server.py
-    chia/seeder/util/service_groups.py
-    chia/seeder/util/service.py
     chia/server/address_manager_sqlite_store.py
     chia/server/address_manager_store.py
     chia/server/introducer_peers.py
@@ -172,13 +169,13 @@ extend_skip=
     tests/core/daemon/test_daemon.py
     tests/core/full_node/full_sync/test_full_sync.py
     tests/core/full_node/ram_db.py
+    tests/core/full_node/stores/test_block_store.py
+    tests/core/full_node/stores/test_coin_store.py
+    tests/core/full_node/stores/test_full_node_store.py
+    tests/core/full_node/stores/test_hint_store.py
     tests/core/full_node/test_block_height_map.py
-    tests/core/full_node/test_block_store.py
-    tests/core/full_node/test_coin_store.py
     tests/core/full_node/test_conditions.py
     tests/core/full_node/test_full_node.py
-    tests/core/full_node/test_full_node_store.py
-    tests/core/full_node/test_hint_store.py
     tests/core/full_node/test_mempool_performance.py
     tests/core/full_node/test_mempool.py
     tests/core/full_node/test_performance.py
@@ -187,7 +184,6 @@ extend_skip=
     tests/core/ssl/test_ssl.py
     tests/core/test_daemon_rpc.py
     tests/core/test_db_conversion.py
-    tests/core/test_farmer_harvester_rpc.py
     tests/core/test_filter.py
     tests/core/test_full_node_rpc.py
     tests/core/util/test_cached_bls.py
