+++ tests/core/full_node/stores/config.py
