+++ tests/core/full_node/stores/test_coin_store.py
