+++ tests/core/full_node/stores/test_full_node_store.py
