+++ tests/core/full_node/stores/test_sync_store.py
